move unit test to separate file
